requirements.txt
discord.py==2.2.2
selenium==4.8.3
tzdata
requests~=2.28.2
ratelimiter~=1.2.0.post0